+++ config/globals.js
@@ -48,7 +48,6 @@ const globals = {
 	SC_SHORT_TITLE: process.env.SC_SHORT_TITLE || 'HPI Schul-Cloud',
 	SMTP_SENDER: process.env.SMTP_SENDER || 'noreply@schul-cloud.org',
 
-	KEEP_ALIVE: process.env.KEEP_ALIVE || false,
 	LEAD_TIME: process.env.LEAD_TIME ? parseInt(process.env.LEAD_TIME, 10) : undefined,
 	/**
 	 * default value 'development' matches default of app.get('env'), but use globals
