+++ src/externalServices/index.js
@@ -1,0 +1,10 @@
+/**
+ * Please add external api services to this location to get an overview over connections.
+ * For any request that is send please use const Api = require('./apiHelper');
+ */
+
+const statusApi = require('./statusApi');
+
+module.exports = {
+	statusApi,
+};
