+++ src/services/fileStorage/proxy-service.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const url = require('url');
 const rp = require('request-promise-native');
+const { Configuration } = require('@hpi-schul-cloud/commons');
 const reqlib = require('app-root-path').require;
 
 const { Forbidden, NotFound, BadRequest, GeneralError } = reqlib('src/errors');
@@ -27,7 +28,6 @@ const { teamsModel } = require('../teams/model');
 const { sortRoles } = require('../role/utils/rolesHelper');
 const { userModel } = require('../user/model');
 const logger = require('../../logger');
-const { KEEP_ALIVE } = require('../../../config/globals');
 const { equal: equalIds } = require('../../helper/compare').ObjectId;
 const {
 	FILE_PREVIEW_SERVICE_URI,
@@ -783,7 +783,7 @@ const newFileService = {
 			)
 			.then((signedUrl) => {
 				const headers = signedUrl.header;
-				if (KEEP_ALIVE) {
+				if (Configuration.get('KEEP_ALIVE')) {
 					headers.Connection = 'Keep-Alive';
 				}
 				return rp({
